add job regular generate (#55)

+++ scheduler/src/provider/scanner.rs
@@ -53,7 +53,10 @@ impl ProviderScanner {
             self.update_providers().await;
             //Update provider map
             self.get_provider_map().await;
-            sleep(Duration::from_secs(CONFIG.update_provider_list_interval)).await;
+            sleep(Duration::from_secs(
+                CONFIG.update_provider_list_interval as u64,
+            ))
+            .await;
         }
     }
